remove some more warnings
git-svn-id: https://icl.cs.utk.edu/svn/magma/trunk@4804 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ sparse-iter/control/magma_zsolverinfo.cpp
@@ -214,7 +214,7 @@ magma_zsolverinfo( magma_z_solver_par *solver_par,
         }
         printf("#======================================================="
                 "======#\n");
-        printf("# GMRES-(%d) solver summary:\n", solver_par->restart);
+        printf("# GMRES-(%d) solver summary:\n", (int) solver_par->restart);
         printf("#    initial residual: %e\n", solver_par->init_res );
         printf("#    iterations: %4d\n", (int) (solver_par->numiter) );
         printf("#    exact final residual: %e\n#    runtime: %.4f sec\n",
